+++ include/CAM/seca.h
@@ -1,7 +1,7 @@
 #ifndef __CAM_SECA_H__
 #  define __CAM_SECA_H__
 
-int seca_card_init(uchar *, int);
+int seca_card_init(uchar *, ushort);
 int seca_do_ecm(ECM_REQUEST *);
 int seca_do_emm(EMM_PACKET *);
 int seca_card_info();
